feat(init): add storage config for public data lakes during quickstart init
An additional storage location is now added for s3://ldb-public/remote during quickstart init, which allows the user to index public ldb data lakes without creating an s3 configuration. A quickstart init is triggered by called `ldb stage` when no instance is configured, so this storage config is not added if an instance is created by calling `ldb init` explicitly.

Closes #299