fix(add): show more informative auto-indexing message
When an ldb instance is configured with `core.auto_index = false` and the `ldb add` command is given a path containing unindexed data objects, a DataObjectNotFoundError is raised. This message only showed the first missing data object's hash. Now it is more informative, showing the file path of the missing data object and instructions on how to resolve the issue by indexing or changing the config.

closes issue #357