fix(index): determine file modified time correctly
Previously, the mtime for annotation and data object files was determined by looking for an "mtime" or "modified" key in the fs.info result, but some filesystems such as S3FileSystem store mtime in a different key. The fs.modified function is designed to get the right value, so it's used instead.

closes #360