fix(index): incorrect overlapping directories error
When a directory and its parent are both given to the `index` command with the `infer` format, an error should be thrown, but this error was happening if two directories are at the same level and one is a prefix of another. Before we were checking if one path started with another. Now we fix the issue by appending a path separator to the end of each path before checking if one starts with another.

closes #386