fix(index): fix storage location check
This fixes an issue where path might be considered outside of a storage location when it's not. If `ldb.fs.posix_path.isin` is called on paths where either one ends with a path separator, then it returns False, because it compares the commonprefix of the two paths against the second path.

closes #443